+++ src/Command/IndexCommand.php
@@ -2,14 +2,20 @@
 
 namespace Algolia\SearchBundle\Command;
 
-
 use Algolia\SearchBundle\IndexManagerInterface;
-use Symfony\Bundle\FrameworkBundle\Command\ContainerAwareCommand;
+use Psr\Container\ContainerInterface;
+use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
+use Symfony\Component\DependencyInjection\ContainerAwareInterface;
+use Symfony\Component\DependencyInjection\ContainerAwareTrait;
+use const E_USER_DEPRECATED;
+use function trigger_error;
 
-abstract class IndexCommand extends ContainerAwareCommand
+abstract class IndexCommand extends Command implements ContainerAwareInterface
 {
+    use ContainerAwareTrait;
+
     protected $indexManager;
 
     public function __construct(IndexManagerInterface $indexManager)
@@ -17,6 +23,19 @@ abstract class IndexCommand extends ContainerAwareCommand
         $this->indexManager = $indexManager;
 
         parent::__construct();
+    }
+
+    /**
+     * @return ContainerInterface
+     */
+    protected function getContainer()
+    {
+        @trigger_error(
+            sprintf('The %s method is deprecated and should not be used. Please wire your dependencies explicitly.', __METHOD__),
+            E_USER_DEPRECATED
+        );
+
+        return $this->container;
     }
 
     protected function getEntitiesFromArgs(InputInterface $input, OutputInterface $output)