+++ src/Command/SearchImportCommand.php
@@ -2,15 +2,33 @@
 
 namespace Algolia\SearchBundle\Command;
 
-
+use Algolia\SearchBundle\IndexManagerInterface;
+use Doctrine\Common\Persistence\ManagerRegistry;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
+use const E_USER_DEPRECATED;
+use function trigger_error;
 
 class SearchImportCommand extends IndexCommand
 {
     protected static $defaultName = 'search:import';
+
+    /**
+     * @var ManagerRegistry|null
+     */
+    private $managerRegistry;
+
+    public function __construct(IndexManagerInterface $indexManager, ManagerRegistry $managerRegistry = null)
+    {
+        parent::__construct($indexManager);
+
+        $this->managerRegistry = $managerRegistry;
+        if ($managerRegistry === null) {
+            @trigger_error('Instantiating the SearchImportCommand without a manager registry is deprecated', E_USER_DEPRECATED);
+        }
+    }
 
     protected function configure()
     {
@@ -27,12 +45,11 @@ class SearchImportCommand extends IndexCommand
 
     protected function execute(InputInterface $input, OutputInterface $output)
     {
-        $doctrine = $this->getContainer()->get('doctrine');
         $entitiesToIndex = $this->getEntitiesFromArgs($input, $output);
         $config = $this->indexManager->getConfiguration();
 
         foreach ($entitiesToIndex as $indexName => $entityClassName) {
-            $manager = $doctrine->getManagerForClass($entityClassName);
+            $manager = $this->getManagerRegistry()->getManagerForClass($entityClassName);
             $repository = $manager->getRepository($entityClassName);
 
             $page = 0;
@@ -62,5 +79,17 @@ class SearchImportCommand extends IndexCommand
 
 
         $output->writeln('<info>Done!</info>');
+    }
+
+    /**
+     * @return ManagerRegistry
+     */
+    private function getManagerRegistry()
+    {
+        if ($this->managerRegistry === null) {
+            $this->managerRegistry = $this->container->get('doctrine');
+        }
+
+        return $this->managerRegistry;
     }
 }